Separated out all PerfettoSqlEvents into their own files
