+++ PerfettoProcessor/Events/PerfettoSqlEvent.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-using System;
 using Perfetto.Protos;
 
 namespace PerfettoProcessor
@@ -49,428 +48,5 @@ namespace PerfettoProcessor
         /// </summary>
         /// <returns></returns>
         public abstract string GetSqlQuery();
-    }
-
-    public class PerfettoSliceEvent : PerfettoSqlEvent
-    {
-        public const string Key = "PerfettoSliceEvent";
-
-        public static string SqlQuery = "select ts, dur, arg_set_id, track_id, name, type, category from slice order by ts";
-        public string Name { get; set; }
-        public string Type { get; set; }
-        public long Duration { get; set; }
-        public long ArgSetId { get; set; }
-        public long Timestamp { get; set; }
-        public string Category { get; set; }
-        public long TrackId { get; set; }
-
-        public PerfettoSliceEvent()
-        {
-
-        }
-
-        public override string GetSqlQuery()
-        {
-            return SqlQuery;
-        }
-
-        public override void ProcessCell(string colName,
-            QueryResult.Types.CellsBatch.Types.CellType cellType,
-            QueryResult.Types.CellsBatch batch,
-            string[] stringCells,
-            CellCounters counters)
-        {
-            var col = colName.ToLower();
-            switch (cellType)
-            {
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellInvalid:
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellNull:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellVarint:
-                    var longVal = batch.VarintCells[counters.IntCounter++];
-                    switch (col)
-                    {
-                        case "ts":
-                            Timestamp = longVal;
-                            break;
-                        case "dur":
-                            Duration = longVal;
-                            break;
-                        case "arg_set_id":
-                            ArgSetId = longVal;
-                            break;
-                        case "track_id":
-                            TrackId = longVal;
-                            break;
-                    }
-
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellFloat64:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellString:
-                    var strVal = stringCells[counters.StringCounter++];
-                    switch (col)
-                    {
-                        case "name":
-                            Name = strVal;
-                            break;
-                        case "type":
-                            Type = strVal;
-                            break;
-                        case "category":
-                            Category = strVal;
-                            break;
-                    }
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellBlob:
-                    break;
-                default:
-                    throw new Exception("Unexpected CellType");
-            }
-        }
-    }
-
-    public class PerfettoArgEvent : PerfettoSqlEvent
-    {
-        public const string Key = "PerfettoArgEvent";
-
-        public static string SqlQuery = "select arg_set_id, flat_key, key, int_value, string_value, real_value, value_type from args order by arg_set_id";
-        public long ArgSetId { get; set; }
-        public string Flatkey { get; set; }
-        public string ArgKey { get; set; }
-        public long IntValue { get; set; }
-        public string StringValue { get; set; }
-        public double RealValue { get; set; }
-        public string ValueType { get; set; }
-
-        public PerfettoArgEvent()
-        {
-
-        }
-
-        public override string GetSqlQuery()
-        {
-            return SqlQuery;
-        }
-
-        public override void ProcessCell(string colName,
-            QueryResult.Types.CellsBatch.Types.CellType cellType,
-            QueryResult.Types.CellsBatch batch,
-            string[] stringCells,
-            CellCounters counters)
-        {
-            var col = colName.ToLower();
-            switch (cellType)
-            {
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellInvalid:
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellNull:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellVarint:
-                    var longVal = batch.VarintCells[counters.IntCounter++];
-                    switch (col)
-                    {
-                        case "arg_set_id":
-                            ArgSetId = longVal;
-                            break;
-                        case "int_value":
-                            IntValue = longVal;
-                            break;
-                    }
-
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellFloat64:
-                    var floatVal = batch.Float64Cells[counters.FloatCounter++];
-                    if (col == "real_value")
-                    {
-                        RealValue = floatVal;
-                    }
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellString:
-                    var strVal = stringCells[counters.StringCounter++];
-                    switch (col)
-                    {
-                        case "flat_key":
-                            Flatkey = strVal;
-                            break;
-                        case "key":
-                            ArgKey = strVal;
-                            break;
-                        case "string_value":
-                            StringValue = strVal;
-                            break;
-                        case "value_type":
-                            ValueType = strVal;
-                            break;
-                    }
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellBlob:
-                    break;
-                default:
-                    throw new Exception("Unexpected CellType");
-            }
-        }
-    }
-
-    public class PerfettoThreadTrackEvent : PerfettoSqlEvent
-    {
-        public const string Key = "PerfettoThreadTrackEvent";
-
-        public static string SqlQuery = "select id, type, name, source_arg_set_id, utid from thread_track";
-        public long ArgSetId { get; set; }
-        public long Id { get; set; }
-        public string Type { get; set; }
-        public string Name { get; set; }
-        public long SourceArgSetId { get; set; }
-        public long Utid { get; set; }
-
-        public PerfettoThreadTrackEvent()
-        {
-
-        }
-
-        public override string GetSqlQuery()
-        {
-            return SqlQuery;
-        }
-
-        public override void ProcessCell(string colName,
-            QueryResult.Types.CellsBatch.Types.CellType cellType,
-            QueryResult.Types.CellsBatch batch,
-            string[] stringCells,
-            CellCounters counters)
-        {
-            var col = colName.ToLower();
-            switch (cellType)
-            {
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellInvalid:
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellNull:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellVarint:
-                    var longVal = batch.VarintCells[counters.IntCounter++];
-                    switch (col)
-                    {
-                        case "arg_set_id":
-                            ArgSetId = longVal;
-                            break;
-                        case "id":
-                            Id = longVal;
-                            break;
-                        case "utid":
-                            Utid = longVal;
-                            break;
-                        case "source_arg_set_id":
-                            SourceArgSetId = longVal;
-                            break;
-                    }
-
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellFloat64:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellString:
-                    var strVal = stringCells[counters.StringCounter++];
-                    switch (col)
-                    {
-                        case "type":
-                            Type = strVal;
-                            break;
-                        case "name":
-                            Name = strVal;
-                            break;
-                    }
-
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellBlob:
-                    break;
-                default:
-                    throw new Exception("Unexpected CellType");
-            }
-        }
-    }
-
-    public class PerfettoThreadEvent : PerfettoSqlEvent
-    {
-        public const string Key = "PerfettoThreadEvent";
-
-        public static string SqlQuery = "select utid, id, type, tid, name, start_ts, end_ts, upid, is_main_thread from thread";
-        public long Utid { get; set; }
-        public long Id { get; set; }
-        public string Type { get; set; }
-        public long Tid { get; set; }
-        public string Name{ get; set; }
-        public long StartTimestamp { get; set; }
-        public long EndTimestamp { get; set; }
-        public long Upid { get; set; }
-        public long IsMainThread{ get; set; }
-
-        public PerfettoThreadEvent()
-        {
-
-        }
-
-        public override string GetSqlQuery()
-        {
-            return SqlQuery;
-        }
-
-        public override void ProcessCell(string colName,
-            QueryResult.Types.CellsBatch.Types.CellType cellType,
-            QueryResult.Types.CellsBatch batch,
-            string[] stringCells,
-            CellCounters counters)
-        {
-            var col = colName.ToLower();
-            switch (cellType)
-            {
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellInvalid:
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellNull:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellVarint:
-                    var longVal = batch.VarintCells[counters.IntCounter++];
-                    switch (col)
-                    {
-                        case "utid":
-                            Utid = longVal;
-                            break;
-                        case "id":
-                            Id = longVal;
-                            break;
-                        case "upid":
-                            Upid = longVal;
-                            break;
-                        case "tid":
-                            Tid = longVal;
-                            break;
-                        case "is_main_thread":
-                            IsMainThread = longVal;
-                            break;
-                        case "start_ts":
-                            StartTimestamp = longVal;
-                            break;
-                        case "end_ts":
-                            EndTimestamp = longVal;
-                            break;
-                    }
-
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellFloat64:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellString:
-                    var strVal = stringCells[counters.StringCounter++];
-                    switch (col)
-                    {
-                        case "type":
-                            Type = strVal;
-                            break;
-                        case "name":
-                            Name = strVal;
-                            break;
-                    }
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellBlob:
-                    break;
-                default:
-                    throw new Exception("Unexpected CellType");
-            }
-        }
-    }
-
-    public class PerfettoProcessEvent : PerfettoSqlEvent
-    {
-        public const string Key = "PerfettoProcessEvent";
-
-        public static string SqlQuery = "select upid, id, type, pid, name, start_ts, end_ts, parent_upid, uid, android_appid, cmdline, arg_set_id from process";
-        public long Upid { get; set; }
-        public long Id { get; set; }
-        public string Type { get; set; }
-        public long Pid { get; set; }
-        public string Name { get; set; }
-        public long StartTimestamp { get; set; }
-        public long EndTimestamp{ get; set; }
-        public long ParentUpid { get; set; }
-        public long Uid { get; set; }
-        public long AndroidAppId { get; set; }
-        public string CmdLine { get; set; }
-        public long ArgSetId { get; set; }
-
-        public PerfettoProcessEvent()
-        {
-
-        }
-
-        public override string GetSqlQuery()
-        {
-            return SqlQuery;
-        }
-
-        public override void ProcessCell(string colName,
-            QueryResult.Types.CellsBatch.Types.CellType cellType,
-            QueryResult.Types.CellsBatch batch,
-            string[] stringCells,
-            CellCounters counters)
-        {
-            var col = colName.ToLower();
-            switch (cellType)
-            {
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellInvalid:
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellNull:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellVarint:
-                    var longVal = batch.VarintCells[counters.IntCounter++];
-                    switch (col)
-                    {
-                        case "upid":
-                            Upid = longVal;
-                            break;
-                        case "id":
-                            Id = longVal;
-                            break;
-                        case "pid":
-                            Pid = longVal;
-                            break;
-                        case "uid":
-                            Uid = longVal;
-                            break;
-                        case "parent_upid":
-                            ParentUpid = longVal;
-                            break;
-                        case "android_appid":
-                            AndroidAppId = longVal;
-                            break;
-                        case "arg_set_id":
-                            ArgSetId = longVal;
-                            break;
-                        case "start_ts":
-                            StartTimestamp = longVal;
-                            break;
-                        case "end_ts":
-                            EndTimestamp = longVal;
-                            break;
-                    }
-
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellFloat64:
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellString:
-                    var strVal = stringCells[counters.StringCounter++];
-                    switch (col)
-                    {
-                        case "type":
-                            Type = strVal;
-                            break;
-                        case "cmdline":
-                            CmdLine = strVal;
-                            break;
-                        case "name":
-                            Name = strVal;
-                            break;
-                    }
-
-                    break;
-                case Perfetto.Protos.QueryResult.Types.CellsBatch.Types.CellType.CellBlob:
-                    break;
-                default:
-                    throw new Exception("Unexpected CellType");
-            }
-        }
     }
 }
